fix: Abandoned cart failing due to unknown field in payload
Abandoned cart payload population uses a function that maps user selected configuration field to their relevant payload values. When an unknown field was used to call the function then the method fails as there is no mapping for that key. This PR will improve the function by skipping such values that are not found instead of failing with PHP error.